# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auron Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: salespersonId
          in: query
          required: false
          schema:
            type: number
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: number
            default: 100
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /finance/summary:
    get:
      operationId: getFinanceSummary
      tags: [finance]
      summary: Financial summary (P&L, EBITDA, net profit)
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: number
        - name: month
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Finance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSummary"
  /finance/expenses:
    get:
      operationId: listOperatingExpenses
      tags: [finance]
      summary: List operating expenses
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: number
        - name: month
          in: query
          required: false
          schema:
            type: number
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operating expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperatingExpense"
    post:
      operationId: createOperatingExpense
      tags: [finance]
      summary: Add an operating expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatingExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatingExpense"
  /finance/expenses/{id}:
    patch:
      operationId: updateOperatingExpense
      tags: [finance]
      summary: Update an operating expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatingExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatingExpense"
    delete:
      operationId: deleteOperatingExpense
      tags: [finance]
      summary: Delete an operating expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Lead:
      type: object
      required: [id, contactName, status, createdAt]
      properties:
        id:
          type: number
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        contactName:
          type: string
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        expectedValue:
          type: ["number", "null"]
        expectedProfit:
          type: ["number", "null"]
        probability:
          type: ["integer", "null"]
        salespersonId:
          type: ["integer", "null"]
        salespersonName:
          type: ["string", "null"]
        dateReceived:
          type: ["string", "null"]
          format: date
        followUpDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [new, contacted, qualified, requirement_received, proposal_sent, negotiation, won, lost]
        lostReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isDemo:
          type: boolean
        weightedValue:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    LeadInput:
      type: object
      required: [contactName, status]
      properties:
        clientId:
          type: number
        contactName:
          type: string
          minLength: 1
        contactPhone:
          type: string
        contactEmail:
          type: string
          
        source:
          type: string
        eventType:
          type: string
        expectedValue:
          type: number
        expectedProfit:
          type: number
        probability:
          type: number
          minimum: 0
          maximum: 100
        salespersonId:
          type: number
        dateReceived:
          type: string
          format: date
        followUpDate:
          type: string
          format: date
        status:
          type: string
          enum: [new, contacted, qualified, requirement_received, proposal_sent, negotiation, won, lost]
        lostReason:
          type: string
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        clientId:
          type: number
        contactName:
          type: string
          minLength: 1
        contactPhone:
          type: string
        contactEmail:
          type: string
          
        source:
          type: string
        eventType:
          type: string
        expectedValue:
          type: number
        expectedProfit:
          type: number
        probability:
          type: number
          minimum: 0
          maximum: 100
        salespersonId:
          type: number
        dateReceived:
          type: string
          format: date
        followUpDate:
          type: string
          format: date
        status:
          type: string
          enum: [new, contacted, qualified, requirement_received, proposal_sent, negotiation, won, lost]
        lostReason:
          type: string
        notes:
          type: string
    LeadListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
...
    FinanceSummary:
      type: object
      required: [revenue, directCosts, grossProfit, grossMarginPct, operatingExpenses, ebitda, ebitdaMarginPct, netProfit, netMarginPct]
      properties:
        revenue:
          type: number
        directCosts:
          type: number
        grossProfit:
          type: number
        grossMarginPct:
          type: number
        operatingExpenses:
          type: number
        ebitda:
          type: number
        ebitdaMarginPct:
          type: number
        netProfit:
          type: number
        netMarginPct:
          type: number
        totalReceivables:
          type: number
        overdueReceivables:
          type: number
    OperatingExpense:
      type: object
      required: [id, category, description, amount, year, month, createdAt]
      properties:
        id:
          type: number
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        gst:
          type: number
        year:
          type: number
        month:
          type: number
        date:
          type: ["string", "null"]
          format: date
        referenceNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OperatingExpenseInput:
      type: object
      required: [category, description, amount, year, month]
      properties:
        category:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        amount:
          type: number
        gst:
          type: number
        year:
          type: number
        month:
          type: number
        date:
          type: string
          format: date
        referenceNumber:
          type: string
    OperatingExpenseUpdate:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        description:
          type: string
        amount:
          type: number
        gst:
          type: number
        year:
          type: number
        month:
          type: number
        date:
          type: string
          format: date
        referenceNumber:
          type: string
...