# lib/api-zod/src/generated/types/operatingExpense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Auron Business OS API
 * OpenAPI spec version: 0.1.0
 */

export interface OperatingExpense {
  id: number;
  category: string;
  description: string;
  amount: number;
  gst?: number;
  year: number;
  month: number;
  /** @nullable */
  date?: Date | null;
  /** @nullable */
  referenceNumber?: string | null;
  createdAt: Date;
}